requirements.txt
llama-index
spacy
python-dotenv
sec_edgar_api
cache_decorator 
html2text